python/20211223/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-bigquery==1.24.0
python-dotenv==0.15.0